# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0",
        "title": "Login/sign-up and logout RESTful API's with node.js using jwt-authentication",
        "description": "Login/sign-up and logout RESTful API's using swagger with mongoose.",
        "contact": {
            "name": "Sarvesh Upadhyay",
            "email": "devtesting676@gmail.com",
            "url": "http://localhost:8000/api-docs"
        }
    },
    "schemes": [
        "http",
        "https"
    ],
    "produces": [
        "application/json"
    ],
    "basePath": "/",
    "apis": [
        "./index.js"
    ],
    "host": "localhost:8000",
    "paths": {
        "/api/get": {
            "get": {
                "tags": [
                    "RESTful API's"
                ],
                "summary": "Get all users",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/Get_User_Responses"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/register": {
            "parameters": [
                {
                    "name": "name",
                    "in": "body",
                    "description": "User with new values of properties",
                    "schema": {
                        "$ref": "#/definitions/Create_User"
                    }
                }
            ],
            "produces": [
                "application/json"
            ],
            "post": {
                "tags": [
                    "RESTful API's"
                ],
                "summary": "Register new user",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/createUserResponses"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/login": {
            "parameters": [
                {
                    "name": "name",
                    "in": "body",
                    "description": "User with new values of properties",
                    "schema": {
                        "$ref": "#/definitions/User_Login"
                    }
                }
            ],
            "produces": [
                "application/json"
            ],
            "post": {
                "tags": [
                    "RESTful API's"
                ],
                "summary": "Login users",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/User_Login_Responses"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/profile": {
            "parameters": [
                {
                    "name": "name",
                    "in": "query",
                    "required": "true",
                    "description": "User with new values of properties"
                },
                {
                    "name": "_id",
                    "in": "query",
                    "required": "true",
                    "description": "User with new values of properties"
                },
                {
                    "name": "email",
                    "in": "query",
                    "required": "true",
                    "description": "User with new values of properties"
                },
                {
                    "name": "firstname",
                    "in": "query",
                    "required": "true",
                    "description": "User with new values of properties"
                },
                {
                    "name": "lastname",
                    "in": "query",
                    "required": "true",
                    "description": "User with new values of properties"
                }
            ],
            "produces": [
                "application/json"
            ],
            "get": {
                "tags": [
                    "RESTful API's"
                ],
                "summary": "Chcek user profile",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/User_profile_Responses"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/logout": {
            "get": {
                "tags": [
                    "RESTful API's"
                ],
                "summary": "Logged-out user",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/Get_Logout_Responses"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "Get_User_Responses": {
            "properties": {
                "Welcome": {
                    "type": "string"
                }
            }
        },
        "Create_User": {
            "properties": {
                "name": {
                    "type": "string"
                },
                "firstname": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "number"
                },
                "password2": {
                    "type": "number"
                }
            }
        },
        "createUserResponses": {
            "properties": {
                "success": {
                    "type": "boolean"
                },
                "user": {
                    "type": "object",
                    "properties": {
                        "name": {
                            "type": "string"
                        },
                        "firstname": {
                            "type": "string"
                        },
                        "lastname": {
                            "type": "string"
                        },
                        "email": {
                            "type": "string"
                        },
                        "password": {
                            "type": "integer"
                        },
                        "password2": {
                            "type": "integer"
                        },
                        "_id": {
                            "type": "integer"
                        }
                    }
                }
            }
        },
        "User_Login": {
            "properties": {
                "name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "password2": {
                    "type": "string"
                }
            }
        },
        "User_Login_Responses": {
            "properties": {
                "isAuth": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "User_profile_Responses": {
            "properties": {
                "isAuth": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "Get_Logout_Responses": {
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}